aplicado CRUD a las paginas, funcional excepto borrar

+++ PagVenta/views.py
@@ -4,7 +4,7 @@ from django.urls import reverse_lazy
 from django.views.generic import ListView, CreateView, UpdateView, DeleteView
 from django import forms
 #-----importamos nuestras estructuras-------------------
-from .models  import Producto
+from .models import Producto
 from .forms import ProductoForm
 
 
@@ -23,35 +23,46 @@ def formulario(request):
 #----CREAMOS UNA FUNCIUON PARA AGREGAR CARRERA----
 
 def inscribir_producto(request):
-    if request.method == "POST":
-        form = ProductoForm(request.POST)
-        if(form.is_valid):
-            model_instance = form.save(commit=False)
-            model_instance.save()
-            return redirect('/addProducto')
+    user = request.user
+    if user.has_perm('app.admin'):
+        if request.method == "POST":
+            form = ProductoForm(request.POST)
+            if(form.is_valid):
+                model_instance = form.save(commit=False)
+                model_instance.save()
+                return redirect('/agregarProducto')
+        else:
+            form = ProductoForm()
+            return render(request, "app/ins_producto.html",{'form':form})
     else:
-        form = ProductoForm()
-        return render(request, "app/ins_carrera.html",{'form':form})
+        return render (request, 'app/error.html')
 
 def listar_producto(request):
-    Producto = Producto.objects.all()
-    return render(request,"app/listar_prducto.html"),{'producto':Producto}
+    productos = Producto.objects.all()
+    return render(request,"app/listar_productos.html",{'productos': productos})
 
-def editar_producto(request, codigo):
-    instancia = Producto.objects.get(id=codigo)
-    form = ProductoForm(instance=instancia)
+def editar_producto (request, producto_id):
+    user = request.user
+    if user.has_perm('app.admin'):
+        instancia = Producto.objects.get(id=producto_id)
+        form = ProductoForm(instance=instancia)
 
-    if request.method == "POST":
-        form = ProductoForm(request.POST, instance= instancia)
-        if form.is_valid():
-            instancia= form.save(commit=False)
-            instancia.save()
+        if request.method == "POST":
+            form = ProductoForm(request.POST, instance= instancia)
+            if form.is_valid():
+                instancia= form.save(commit=False)
+                instancia.save()
 
-    return render(request, "app/editar_producto.html",{'form':form})
+        return render(request, "app/editar_producto.html",{'form':form})
+    else:
+        return render (request, 'app/error.html')
 
-def borrar_producto(request, codigo):
-    instancia = Producto.objects.get(id=codigo)
-    instancia.delete()
-    return redirect("/")
 
-    
+def borrar_producto(request, producto_id):
+    user = request.user
+    if user.has_perm('app.admin'):
+        instancia = Producto.objects.get(id=producto_id)
+        instancia.delete()
+        return redirect("/")
+    else:
+        return render (request, 'app/error.html')
